+++ packages/protons/bin/protons.ts
@@ -1,7 +1,6 @@
 #! /usr/bin/env node
 
 import meow from 'meow'
-
 import { generate } from '../src/index.js'
 
 async function main (): Promise<void> {
@@ -15,11 +14,12 @@ async function main (): Promise<void> {
   Examples
     $ protons ./path/to/file.proto ./path/to/other/file.proto
 `, {
+    // @ts-expect-error wrong version is hoisted?!
     importMeta: import.meta,
     flags: {
       output: {
         type: 'string',
-        alias: 'o'
+        shortFlag: 'o'
       }
     }
   })