+++ packages/protons/test/unsupported.spec.ts
@@ -1,7 +1,6 @@
 /* eslint-env mocha */
 
 import { expect } from 'aegir/chai'
-
 import { generate } from '../src/index.js'
 
 describe('unsupported', () => {